requirements.txt
Django==3.0.7
djangorestframework==3.11.0
docutils==0.16
drf-yasg==1.17.1
drfdocs==0.0.11
requests==2.23.0
psycopg2==2.8.5
urllib3==1.25.9